+++ CHANGELOG.md
@@ -11,6 +11,10 @@ Every entry has a category for which we use the following visual abbreviations:
 
 ## Unreleased
 
+- 🎁 VAST now merges the contents of all used configuration files instead of
+  using only the most user-specific file.
+  [#1040](https://github.com/tenzir/vast/pull/1040)
+
 - 🎁 The output of the `status` command was restructured with a strong focus on
   usability. The new flags `--detailed` and `--debug` add additional content to
   the output. [#995](https://github.com/tenzir/vast/pull/995)